pyproject.toml
[project]
name = "glance"
version = "0.1.0"
description = "Get LLM summaries of YouTube videos and Reddit threads from your terminal"
readme = "README.md"
authors = [
    { name = "wrinkledeth", email = "wrinkled.eth@gmail.com" }
]
requires-python = ">=3.12"
dependencies = [
    "anthropic>=0.52.0",
    "httpx>=0.28.0",
    "python-dotenv>=1.0.0",
]

[project.scripts]
glance = "glance.cli:main"

[build-system]
requires = ["uv_build>=0.11.3,<0.12.0"]
build-backend = "uv_build"
